pyproject.toml
[project]
name = "kavak-agent"
version = "0.1.0"
description = "Bot comercial de Kavak con integración WhatsApp y LLM"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "openai>=2.13.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pgvector>=0.3.0",
    "twilio>=9.0.0",
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "thefuzz>=0.22.0",
    "python-multipart>=0.0.6",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.5",
    "langgraph>=0.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
